Refactoring - no functional change. Introducing SecurityUtils.
